+++ web/themes/custom/itkdev/itkdev_project_theme/js/project-scripts.js
@@ -1,1 +1,2 @@
 // add scripts specific to the project here
+require("./selectColorBlock");