+++ web/themes/custom/itkdev/itkdev_project_theme/tailwind.config.js
@@ -1,0 +1,15 @@
+/**
+ * This file is merged into itkdev_base_theme tailwind config.
+ * This file is required by itkdev_base_theme.
+ *
+ * Project specific tailwind config should be added to this file.
+ */
+
+module.exports = {
+  safelist: [
+    "font-bold",
+    "bg-slate-300",
+    "bg-yellow-700",
+    "bg-red-700"
+  ],
+};